Removed tests refactoring which will be included as a seperate PR
