+++ modelservice/src/test/java/com/expedia/adaptivealerting/modelservice/test/ObjectMother.java
@@ -20,6 +20,13 @@ import com.expedia.adaptivealerting.modelservice.domain.mapping.Expression;
 import com.expedia.adaptivealerting.modelservice.domain.mapping.Field;
 import com.expedia.adaptivealerting.modelservice.domain.mapping.Operand;
 import com.expedia.adaptivealerting.modelservice.domain.mapping.Operator;
+import com.expedia.adaptivealerting.modelservice.domain.mapping.User;
+import com.expedia.adaptivealerting.modelservice.domain.mapping.ConsumerDetectorMapping;
+import com.expedia.adaptivealerting.modelservice.domain.percolator.BoolCondition;
+import com.expedia.adaptivealerting.modelservice.domain.percolator.MustCondition;
+import com.expedia.adaptivealerting.modelservice.domain.percolator.PercolatorDetectorMapping;
+import com.expedia.adaptivealerting.modelservice.domain.percolator.Query;
+import com.expedia.adaptivealerting.modelservice.metricsource.graphite.GraphiteResult;
 import com.expedia.adaptivealerting.modelservice.web.request.AnomalyRequest;
 import com.expedia.adaptivealerting.modelservice.metricsource.MetricSourceResult;
 import com.expedia.adaptivealerting.modelservice.util.DateUtil;
@@ -34,13 +41,23 @@ import java.util.Map;
 import java.util.UUID;
 
 public class ObjectMother {
-
     private static final ObjectMother MOM = new ObjectMother();
+
     public static ObjectMother instance() {
         return MOM;
     }
 
     private ObjectMother() {
+    }
+
+    public GraphiteResult[] getGraphiteData() {
+        GraphiteResult[] results = new GraphiteResult[1];
+        GraphiteResult result = new GraphiteResult();
+        result.setDatapoints(getDataPoints());
+        result.setTags(getTags());
+        result.setTarget("target");
+        results[0] = result;
+        return results;
     }
 
     public MetricSourceResult getMetricData() {
@@ -71,18 +88,39 @@ public class ObjectMother {
      *
      * @return Detector document with the UUID set to {@literal null}.
      */
-    public DetectorDocument buildDetectorDocument() {
-        return new DetectorDocument()
-                .setCreatedBy("user")
-                .setType("constant-detector")
-                .setConfig(buildDetectorConfig());
+    public DetectorDocument getDetectorDocument() {
+        val detector = new DetectorDocument();
+        detector.setCreatedBy("user");
+        detector.setType("constant-detector");
+
+        Map<String, Object> detectorConfig = new HashMap<>();
+        val thresholds = "{\"thresholds\": {\"lowerStrong\": \"90\", \"lowerWeak\": \"70\"}}";
+        val detectorParams = toObject(thresholds);
+        detectorParams.put("type", "LEFT_TAILED");
+        detectorConfig.put("params", getDetectorParams());
+        detector.setConfig(detectorConfig);
+        return detector;
     }
 
-    public DetectorDocument buildIllegalParamsDetectorDocument() {
-        return new DetectorDocument()
-                .setCreatedBy("user")
-                .setType("constant-detector")
-                .setConfig(buildIllegalDetectorConfig());
+    public Map<String, Object> getIllegalDetectorParams() {
+        val thresholds = "{\"thresholds\": {\"lowerStrong\": \"90\", \"lowerWeak\": \"70\"}}";
+        val detectorParams = toObject(thresholds);
+        detectorParams.put("type", "LEFT_TAILED");
+        return detectorParams;
+    }
+
+    public DetectorDocument getIllegalParamsDetector() {
+        DetectorDocument detector = new DetectorDocument();
+        detector.setCreatedBy("user");
+        detector.setType("constant-detector");
+
+        Map<String, Object> detectorConfig = new HashMap<>();
+        val thresholds = "{\"thresholds\": {\"lowerStrong\": \"90\", \"lowerWeak\": \"70\"}}";
+        val detectorParams = toObject(thresholds);
+        detectorParams.put("type", "LEFT_TAILED");
+        detectorConfig.put("params", getIllegalDetectorParams());
+        detector.setConfig(detectorConfig);
+        return detector;
     }
 
     public DetectorDocument getElasticsearchDetector() {
@@ -93,6 +131,27 @@ public class ObjectMother {
                 .setEnabled(true)
                 .setTrusted(true)
                 .setLastUpdateTimestamp(DateUtil.toUtcDate("2019-04-06 22:00:00"));
+    }
+
+    public PercolatorDetectorMapping getPercolatorDetectorMapping() {
+        PercolatorDetectorMapping percolatorDetectorMapping = new PercolatorDetectorMapping();
+        Query query = new Query();
+        BoolCondition boolCondition = new BoolCondition();
+        List<MustCondition> mustConditions = new ArrayList<>();
+        MustCondition mustCondition = new MustCondition();
+        Map match = new HashMap<>();
+        match.put("name", "sample-web");
+        mustCondition.setMatch(match);
+        mustConditions.add(mustCondition);
+        boolCondition.setMust(mustConditions);
+        query.setBool(boolCondition);
+        percolatorDetectorMapping.setCreatedTimeInMillis(1554828886);
+        percolatorDetectorMapping.setEnabled(true);
+        percolatorDetectorMapping.setLastModifiedTimeInMillis(1554828886);
+        percolatorDetectorMapping.setUser(new User("test-user"));
+        percolatorDetectorMapping.setConsumerDetectorMapping(buildConsumerDetectorMapping("aeb4d849-847a-45c0-8312-dc0fcf22b639"));
+        percolatorDetectorMapping.setQuery(query);
+        return percolatorDetectorMapping;
     }
 
     public Expression getExpression() {
@@ -106,36 +165,28 @@ public class ObjectMother {
         return expression;
     }
 
-    private Map<String, Object> buildDetectorConfig() {
-        Map<String, Object> detectorConfig = new HashMap<>();
-        detectorConfig.put("params", getLegalDetectorParams());
-        return detectorConfig;
-    }
-
-    private Map<String, Object> buildIllegalDetectorConfig() {
-        Map<String, Object> detectorConfig = new HashMap<>();
-        detectorConfig.put("params", getIllegalDetectorParams());
-        return detectorConfig;
-    }
-
-    private Map<String, Object> getLegalDetectorParams() {
-        // FIXME Use serialization rather than hand-crafted JSON. [WLW]
-        val thresholds = "{\"thresholds\": {\"lowerStrong\": \"70\", \"lowerWeak\": \"90\"}}";
-        val detectorParams = toObject(thresholds);
-        detectorParams.put("type", "LEFT_TAILED");
-        return detectorParams;
-    }
-
-    private Map<String, Object> getIllegalDetectorParams() {
-        val thresholds = "{\"thresholds\": {\"lowerStrong\": \"90\", \"lowerWeak\": \"70\"}}";
-        val detectorParams = toObject(thresholds);
-        detectorParams.put("type", "LEFT_TAILED");
-        return detectorParams;
-    }
-
     @SneakyThrows
     private Map<String, Object> toObject(String message) {
         return new ObjectMapper().readValue(message, HashMap.class);
     }
-}
 
+    private String[][] getDataPoints() {
+        String[][] datapoints = new String[2][2];
+        datapoints[0][0] = String.valueOf(78.0);
+        datapoints[0][1] = String.valueOf(1548829800);
+        datapoints[1][0] = String.valueOf(81.0);
+        datapoints[1][1] = String.valueOf(1548830400);
+        return datapoints;
+    }
+
+    private Map<String, Object> getTags() {
+        Map<String, Object> tags = new HashMap<String, Object>();
+        tags.put("lob", "hotel");
+        tags.put("pos", "expedia-com");
+        return tags;
+    }
+
+    private ConsumerDetectorMapping buildConsumerDetectorMapping(String detectorUuid) {
+        return new ConsumerDetectorMapping("cid", UUID.fromString(detectorUuid));
+    }
+}
